refactor: fold active_leads_count counter into the existing LeadObserver
Team Lead asked for this to live on Lead's default observer instead of a
separate #[ObservedBy] class, matching how the rest of LeadObserver's
side effects are wired (registered once in EventServiceProvider::boot()).
RecalculateActiveLeadsCountCommand stays — a drift-repair tool for the
extreme case a bulk write bypasses the observer, not something run
routinely.

The counter decrement in deleted() runs before the pre-existing (unrelated,
out of scope) getSocialChannel() call, so a real hard delete still applies
it even though that call currently throws.
