+++ src/Domains/Guild/Leads/Models/Lead.php
@@ -9,7 +9,6 @@ use Baka\Traits\DynamicSearchableTrait;
 use Baka\Traits\HasLightHouseCache;
 use Baka\Traits\UuidTrait;
 use Baka\Users\Contracts\UserInterface;
-use Illuminate\Database\Eloquent\Attributes\ObservedBy;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
@@ -25,7 +24,6 @@ use Kanvas\Guild\Leads\Enums\ConfigurationEnum;
 use Kanvas\Guild\Leads\Enums\LeadFilterEnum;
 use Kanvas\Guild\Leads\Enums\LeadGroupStatusEnum;
 use Kanvas\Guild\Leads\Factories\LeadFactory;
-use Kanvas\Guild\Leads\Observers\LeadActiveLeadsCounterObserver;
 use Kanvas\Guild\LeadSubSources\Models\LeadSubSource;
 use Kanvas\Guild\Models\BaseModel;
 use Kanvas\Guild\Organizations\Models\Organization;
@@ -76,7 +74,6 @@ use Throwable;
  * @property string $is_duplicate
  * @property string $third_party_sync_status
  */
-#[ObservedBy(LeadActiveLeadsCounterObserver::class)]
 class Lead extends BaseModel implements EventResourceInterface
 {
     use UuidTrait;
